Slew of minor clang-tidy fixes [WIP]
Make minor linter and style improvements, with a couple of minor parsing
bug fixes found. Some of this is because of recently added code
I also upgraded my local clang-tidy from v20->v22, which introduces
a few new checkers. The most significant newly flagged errors are unbounded
countainer checks i.e. mostly for std::vector. Lastly, I fixed the clang-tidy
CI script to run correctly - the I flag was messing things up

I've started going through and making fixes where appropriate. Almost everywhere
unbounded accesses are warned I can probably fix. In some important places, e.g., in
the model code, I may want to keep the faster unbounded [] access, but maybe I can think
of a cleverer/safer way to do some of these accesses. Some more work is therefore
needed to get back to 0 unsuppressed clang-tidy warnings...

+++ src/core.cpp
@@ -7,6 +7,7 @@
 #include <functional> // For std::reference_wrapper
 #include <sstream>
 #include <stdexcept>
+#include <string>
 #include <vector>
 
 #include "arch.hpp"
@@ -84,9 +85,10 @@ sanafe::PipelineUnit *sanafe::Core::get_hw(const std::string &hw_name,
     if (!hw_found)
     {
         const std::string error("Could not find h/w (with name:" + hw_name +
-                ") that implements synapse:" + std::to_string(is_synapse) +
-                ", dendrite:" + std::to_string(is_dendrite) +
-                ", soma:" + std::to_string(is_soma));
+                ") that implements synapse:" +
+                std::to_string(static_cast<int>(is_synapse)) +
+                ", dendrite:" + std::to_string(static_cast<int>(is_dendrite)) +
+                ", soma:" + std::to_string(static_cast<int>(is_soma)));
         INFO("Error: %s\n", error.c_str());
         throw std::runtime_error(error);
     }
